refactor: 액세스 토큰 이름 상수화
