+++ src/constants/tokens.js
@@ -1,0 +1,3 @@
+export const TOKEN_INFO = Object.freeze({
+  ACCESS_TOKEN_NAME: process.env.REACT_APP_ACCESS_TOKEN_NAME || 'act',
+});
